Codechange: Simplify Pillow imports
Previously, this tried a fallback to `import Image`, but Pillow has
stopped supporting that in version 1.0, so there is no need to keep
carrying that around.

Note that in some places, an ImportError for Pillow was and still is
ignored, since Pillow is not required for all codepaths. See commit
5612c5b2 where this was introduced.