Fix #209: Cargo 'profit' callback applied a unit conversion, when there was no unit.
The 'profit' callback in NML was supposed to be used like this:
  cargo_income = callback_result * cargo_amount * price_factor
But since NML defines 'price_factor' as income per 10 units over 20 tiles without time-penalty (=255/256), this definition results in a very weird meaning of 'callback_result'.
NML applied some unit conversion to the callback result, which was supposed to revert this 20*10*255/256 scaling, but this only made it harder to use.

By removing the unit conversion in the callback, the 'callback_result' can now be defined as relative scaling of the 'price_factor'.
To do something similar as the default-payment, you can now use
  callback_result = distance * time_penalty
with 'time_penalty' in range 0..1.