Fix: Clean up generated parser files
The setup.py file overrides the build step to generate parsetab.py and
lextab.py, but does not clean them up. This adds an override to the
clean step to remove them again.

Note that this uses the distutils version of the command as a basis,
since (unlike the build_py step) setuptools does not provide its own
version of it but just uses the distutils version.